front: ui-charts: move generic types in common
