+++ front/ui/ui-charts/src/chronogram/components/ChronogramCanvas.tsx
@@ -10,9 +10,8 @@ import { useMouseTracking } from '../../common/hooks/useMouseTracking';
 import { useSize } from '../../common/hooks/useSize';
 import { TimeCaptions } from '../../common/layers/TimeCaptions';
 import TimeGraduations from '../../common/layers/TimeGraduations';
-import type { MouseContextType, Point } from '../../common/types';
+import type { MouseContextType, Point, DataPoint } from '../../common/types';
 import { DEFAULT_THEME } from '../../spaceTimeChart/lib/consts';
-import type { DataPoint } from '../../spaceTimeChart/lib/types';
 import { ChronogramContext } from '../lib/context';
 import type { ChronogramContextType, ChronogramCanvasProps } from '../lib/types';
 import { OccupancyBlocksLayer } from './OccupancyBlocksLayer';
