+++ front/ui/ui-charts/src/common/index.ts
@@ -7,11 +7,13 @@ export { getCrispLineCoordinate } from './helpers/time';
 export type {
   CurveOutline,
   CurveStyle,
-  HoveredItem,
+  DataPoint,
   DrawingFunction,
+  HoveredItem,
   PickingDrawingFunction,
   PickingElement,
   Point,
+  PathLevel,
 } from './types';
 
 export { MouseContext } from './context';
